gpu: nvgpu: MISRA 11.3 fixes to gr ctx code
MISRA rule 11.3 states that a cast shall not be performed between
a pointer to object type and a pointer to a different object type.

The following gr context routines convert a gr context buffer pointer
in the form of a u8 * to a u32 * before referencing the context buffer's
contents:

 * gm20b_ctxsw_prog_check_main_image_header_magic
 * gm20b_ctxsw_prog_check_local_header_magic
 * gm20b_ctxsw_prog_get_num_gpcs
 * gm20b_ctxsw_prog_get_num_tpcs
 * gm20b_ctxsw_prog_get_extended_buffer_size_offset
 * gm20b_ctxsw_prog_get_ppc_info
 * gm20b_ctxsw_prog_get_local_priv_register_ctl_offset

Because the gr context buffer pointer starts out as a u32 * the
conversion to a u8 * isn't strictly necessary.

So this patch eliminates the conversion and the 11.3 rule violations
accordingly.

JIRA NVGPU-782

Reviewed-on: https://git-master.nvidia.com/r/2102986
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ drivers/gpu/nvgpu/hal/gr/ctxsw_prog/ctxsw_prog_gm20b.h
@@ -68,14 +68,14 @@ void gm20b_ctxsw_prog_set_priv_access_map_addr(struct gk20a *g,
 	struct nvgpu_mem *ctx_mem, u64 addr);
 void gm20b_ctxsw_prog_disable_verif_features(struct gk20a *g,
 	struct nvgpu_mem *ctx_mem);
-bool gm20b_ctxsw_prog_check_main_image_header_magic(u8 *context);
-bool gm20b_ctxsw_prog_check_local_header_magic(u8 *context);
-u32 gm20b_ctxsw_prog_get_num_gpcs(u8 *context);
-u32 gm20b_ctxsw_prog_get_num_tpcs(u8 *context);
-void gm20b_ctxsw_prog_get_extended_buffer_size_offset(u8 *context,
+bool gm20b_ctxsw_prog_check_main_image_header_magic(u32 *context);
+bool gm20b_ctxsw_prog_check_local_header_magic(u32 *context);
+u32 gm20b_ctxsw_prog_get_num_gpcs(u32 *context);
+u32 gm20b_ctxsw_prog_get_num_tpcs(u32 *context);
+void gm20b_ctxsw_prog_get_extended_buffer_size_offset(u32 *context,
 	u32 *size, u32 *offset);
-void gm20b_ctxsw_prog_get_ppc_info(u8 *context, u32 *num_ppcs, u32 *ppc_mask);
-u32 gm20b_ctxsw_prog_get_local_priv_register_ctl_offset(u8 *context);
+void gm20b_ctxsw_prog_get_ppc_info(u32 *context, u32 *num_ppcs, u32 *ppc_mask);
+u32 gm20b_ctxsw_prog_get_local_priv_register_ctl_offset(u32 *context);
 u32 gm20b_ctxsw_prog_hw_get_ts_tag_invalid_timestamp(void);
 u32 gm20b_ctxsw_prog_hw_get_ts_tag(u64 ts);
 u64 gm20b_ctxsw_prog_hw_record_ts_timestamp(u64 ts);
